// src/app/models/approved-mockups.model.ts
export interface ApprovedMockupModel {
    mockupId: number;
    teamName?: string;
    mainCategoryId?: number;
    categoryId?: number;
    subCategoryId?: number;
    productId?: number;
    fabricTypeId?: number;
    neckStyleId?: number;
    frontDescription?: string;
    backDescription?: string;
    leftSleeveDesc?: string;
    rightSleeveDesc?: string;
  }
  